Fixing some edge cases in cpp00 and cpp01, Updated formatting

+++ cpp00/ex01/PhoneBook.hpp
@@ -6,7 +6,7 @@
 /*   By: charles <charles.cabergs@gmail.com>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/13 09:06:23 by charles           #+#    #+#             */
-/*   Updated: 2020/04/13 09:06:39 by charles          ###   ########.fr       */
+/*   Updated: 2020/11/09 09:28:38 by cacharle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,16 +22,16 @@ class PhoneBook
 {
 public:
 	PhoneBook();
-	void add(Contact contact);
-	Contact* search(std::string needle);
-    size_t getSize() const;
-    Contact const &get(int index) const;
+	void           add(Contact contact);
+	Contact*       search(std::string needle);
+    size_t         getSize() const;
+    Contact const& get(size_t index) const;
 
 private:
 	Contact m_contacts[CONTACTS_SIZE];
-	int m_size;
+	size_t  m_size;
 };
 
-std::ostream &operator<<(std::ostream &out, PhoneBook const &p);
+std::ostream& operator<<(std::ostream& out, PhoneBook const& p);
 
 #endif
